test: add sample data test
Used XSPEC on the same dataset to get fiducial values for the fit
parameters for a simple model. These agree very well with the values
obtained by SpectralFitting.

Made a test case using the sample data to ensure this is never broken.
Includes an implicit test for `regroup!` and `mask_domain!`.