Feat: new OGIP parser and hopefully simpler datasets (#48)
* feat: finish OGIP parsing

All of the component OGIP parsing are now in place and type stable.
Tests suite expanded too, and checking inferrence.

* feat!: reworked and simplified datasets

Datasets now store primitive structures with optionals (i.e.
`Union{Missing,T}`). This seems to quite dramatically simplify much of
the API, but some care needs to be taken with type safety. Nevertheless,
this is an improvement on the old design, where type stability was
almost impossible.

Removed all the old OGIP code, as it is now entirely replaced with the
new parsing.

Added a `SpectralFilePaths` structure for managing the (possibly)
missing filepaths. This should also hopefully make adding new parsers a
little easier.

Added support for reading background spectra, which, similar to the
signal, are simply `Spectrum` types. These are not yet subtracted, but
their units are normalised.

Spectrum units now all force "counts / s" if possible, to avoid things
getting mismatched. The accessors for counts / rate are still
implemented.

Avoid all property based access in favour of method based access. API is
*not yet exported*.

Updated tests for changes, including the dummy datasets. All tests
passing.

* chore: bump version

* fix: regrouping, plotting, and fitting

Regrouping code fixed for the new dataset. Now operates inplace instead
of out of place. `mask_energy!` renamed to `mask_domain!` to reflect a
more general purpose.

Fixed various plotting utility methods that tried to access now
non-existent fields.

Fixed fitting routines that were assembling the fitting problems and
were not using the `target_*` and `domain_*` API.

bug: Fitting on the MCG-6-30-15 test script still seems to be bad, with
a > 1.0.

* fix: printing fixes for background

Background information was not being masked out correctly, showing
negative counts occasionally.

Corrected the printing order and included a bin count.

* feat: result plot stepmid

Plot the results line as piecewise constant instead of a continuous line
to reflect bin boundaries better. This is optional and can be
overwritten with the `seriestype` kwarg to `plot`.

* test: add sample data test

Used XSPEC on the same dataset to get fiducial values for the fit
parameters for a simple model. These agree very well with the values
obtained by SpectralFitting.

Made a test case using the sample data to ensure this is never broken.
Includes an implicit test for `regroup!` and `mask_domain!`.

* fix: model base name resolved and remove show

* test: fix missing model data