fix: add _parse_any function
Since FITS is a strongly-typed data format, sometimes we get annoying
behaviour. Something that is clearly a number is stored in the table as
a string, and the

        `T(value)`

conversion syntax fails, as it requires a `parse` call to convert a
string to a number. This commit adds a `_parse_any` function that
parses or converts a value to a given type, depending on whether it is
an `AbstractString` or not.