feat(data): match channels of spectrum and response
There was an issue with one of the old EXOSAT datasets where there was a
mismatch in the channels in the response matrix and the spectrum, which
lead to the masking selecting the wrong bins when fitting.

This commit adds a check (on by default, but toggleable) that drops
channels in the response matrix that are not in the spectrum.

I've made it optional, as if someone is interested in seeing the full
response matrix they can use the same API but just pass false to the
keyword argument.