feat: units in fitting
Started adding the code that tracks that the model, data, and fit
statistic are used to change the units used when fitting. For example,
when using Cash stat, we want to fit in counts, but for ChiSquared we
want to fit in counts / (s * keV^-1), or similar.

There is a looming issue here related to type instability with the
units, since the units are basically compile time structs, but the way
we store them in e.g. Spectrum is as a union so they may easily be
modified.

This means we can't rely on the dataset units for codegen.

deprecated: normalize!