fix: parameter cache frozen with dual numbers
There was a bug where using dual numbers with a different chunk size
would ruin the alignment of the temporary buffer, which means all the
frozen values we had baked into the array were being read out with a
different stride.

Fix is to use the mask to update the frozen and free parameter together,
so that the values are always correct no matter the chunk size.