feat: model wrappers and AutoCache
Added a new abstract type for creating model wrappers. These are models
which wrap other models without introducing new parameters, but can
change the way the underlying model is used.

One example of this is included: `AutoCache`, which caches the last
output of the model and returns the cached output if the parameters have
all changed by less than some absolute tolerance. This can save the
number of calls to expensive models where the parameter space is not
changing by much, and speedup fits.