fix(AutoCache): ignore normalisation in check
The normalisation constant is applied outside of the model, so if the
model is additive the normalisation parameter does not need to be
checked to see if the model needs re-evaluation. This makes a big
performance difference if fitting simultaneous datasets on the same with
bound parameters domain but with different normalisations.