feat(model): delta function
Added a delta function model, which is in essence an alias to a very
thin Gaussian. This can be at some point updated to a more faithful
implementation, provided we find a good way of propagating gradients as
the line moves through the domain.

Intended use is for emission lines or as the basis for convolutions.