feat(model): AsConvolution wrapper
This wrapper can turn any AbstractSpectralModel into a convolutional
model, and will remove additive normalisations if required.

Convolution function is a modification of the Toeplitz matrix method,
which performs the matrix multiplications without allocating the matrix.
The benefit this has over, e.g., the FFT(W) method of taking
convolutions, in that we can propagate gradients through the operation.