feat: add describe function for FittingProblem
This is to begin addressing some of the parameter binding issues that were
being encountered. The `describe` function differs from just displaying the
FittingProblem in that it will resolve and pretty print the parameter bindings.

Bound parameters will no longer appear as `FREE` or `FROZEN` but rather list
the model and that model's parameter symbol that they are bound to.