fix: parameter binding logic
One of the concepts that made debugging the parameter bindings difficult is
that ultimately the parameter bindings refer to the vector of _free
parameters_, not the immediate parameters of the model. That is, the presence
of frozen parameters could throw off the expected binding resolution.

To fix this, we defer adjusting the parameter indices until we construct the
fit, and introduce a new function `adjust_free_bindings` which will do the
index shuffling and remove frozen parameter bindings.

The bindings that are kept in `prob.bindings` are now the expected model =>
parameter index mappings without the frozen adjustment, which makes reasoning
whether a binding is correct or not that bit easier.