feat!: add Tag parameter to SpectralData
This enormously simplifies derived datatypes, and removed the API
forwarding macro that was unreliable to maintain. Follows from the
Tagging idea introduced with BinnedData, and allows derived types to be
implemented more easily.

Updated OGIP, XMM and NuSTAR datatypes to use this tagging system
instead of adding new types.

Modified test cases to sync with changes.