feat: add table model utilities
For loading in XSPEC-style table models, there are now a number of
utilities to make wrapping models easier.

One of the main difficulties is that SpectralFitting.jl needs to know
how many parameters are in a model at compile time, which cannot be
inferred from the table model file. The workaround for now is to have
the caller pass a `Val{N}` with the number of parameters.

Adds the MultiLinearInterpolations package as a dependency.

Needs this package in order to be able to provide the interpolation
logic needed for table models.

Makes a number of assumptions about what the table models can be like,
many of which are documented, but not asserted.