feat!: major rework of core components
Entirely removed the code paths that made use of generated reflection
routines. Everything is now worked out during the standard Julia
inference stages, which has made compilation times noticeably better,
and even incurred a few performance improvements. Benchmarks to follow.

Composite models descend their call tree in left-then-right order,
instead of what we previously had. This leads to some more natural
orderings e.g. in model printout. Accessing parameters can no-longer be
done by referring to the parameter symbols, but instead hierarchically.
This avoid a number of problems when modifying models but wanting to
keep parameter binding sections.

Reworked the FittingConfig and result types. These are much simpler, and
treat single model as a special case of multi-model, to avoid having to
track the implementation of both of those two things separately.

Parameter binding is also reworked, and is much simpler. Instead of
trying to account for intra-model binds, frozen parameters, redundant
parameters, and so on, all in one go, we just use a view into a
parameter cache which contains both the free and frozen parameters. Some
bugs related to parameter binding have been fixed consequently, and the
`bind!` syntax has changed with the introduction of a `bindall!` to
disambiguate the action of binding.

Not everything is working perfectly. Model printing, for example, is now
different, and there are many API changes and some parts of especially
the simulation and auto caching do not fully work yet.

There is still a lot of dead code that needs to be removed, but that
will be for later commits.