feat: no named parameter tuples + better wrappers
Removed all code that used the `unpack_*` variants that returned named
tuples. There's no real reason to make everything compile time
resolveable, since that is not the bottleneck of model performance, and
a few allocations here and there during setup are not important.

Those types of methods are superceded by `parameter_names` and
`parameter_vector`, which returns the symbols and vector of parameters
respectively.

Part of this also heavily modifies how the `invokmodel` system works. To
illustrate this:

      invokemodel             High level interface, allocates output
                              buffers
           |
           v

      invokemodel!            High level mutating interface, unpacks
                              the parameters, and calls
           |                  `remake_with_parameters` to ensure model
           |                  has correct types.
           |
           v

      _inner_invokemodel!     Used as a sink, so that all entries into
                              model evaluation pass through a single
           |                  function that can be bootstrapped by
           |                  other or future components.
           v

      _invoke_guard!          Applies any transformations, e.g.
                              normalisation folding.
           |
           v

      _invoke!                User defined implementation.

This means various wrappers can add hooks at different points of model
evaluation to modify the behaviour. It would probably be good to
continue working in this direction and eventually rename these functions
so that their meaning is less opaque.

Also removed some dead code.