feat: patched FitParam
FitParam may now be `patched`, which means a ParameterPatch is modifying
it. Since it's not obvious if a parameter patch is merely modifying or
overwrite a value, it doesn't make sense to force a patched parameter to
be frozen. As such, it is still up to the caller to decide whether a
patched parameter is frozen or not. Being patched merely augments the
printout with a `/PATCHED` status, so that it is clear that a parameter
in the context of a model is being modified.